awsv4, basic working
